requirements.txt
cffi==1.15.0
cycler==0.11.0
fonttools==4.28.2
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
PyAudio==0.2.11
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
sounddevice==0.4.3
tomli==1.2.2
docutils==0.18.1
pyserial==3.5
robotframework==4.1.2
robotremoteserver==1.1
